requirements.txt
django
django-allauth
twitter
pytz
